move block filter to versioned
