+++ packages/common/src/project/versioned/v1_0_0/types.ts
@@ -26,10 +26,6 @@ export interface NodeOptions {
   unfinalizedBlocks?: boolean;
 }
 
-export interface FileReference {
-  file: string;
-}
-
 export interface ProjectManifestV1_0_0<T extends object = TemplateBase, D extends object = BaseDataSource>
   extends Omit<ProjectManifestV0_2_1<T, D>, 'network'> {
   dataSources: D[];
@@ -44,4 +40,9 @@ export interface ProjectManifestV1_0_0<T extends object = TemplateBase, D extend
       file: string;
     };
   };
+}
+
+export interface BlockFilter {
+  modulo?: number;
+  timestamp?: string;
 }
